aria-app/requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Executor==0.9.0
Flask-RESTX==1.3.0

# Database
SQLAlchemy==2.0.23
mysqlclient==2.2.0

# Authentication & Security
bcrypt==4.1.2

# Face Recognition & ML
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
keras-facenet==0.3.0

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.1

# Development (optional)
# pytest==7.4.3
# pytest-flask==1.3.0
# black==23.11.0
# ruff==0.1.6

